requirements.txt
asgiref==3.2.10
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1
django-filter==2.4.0
django-gsheets==0.0.10
django-import-export==2.4.0
djangorestframework==3.12.2
et-xmlfile==1.0.1
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
jdcal==1.4.1
Markdown==3.3.3
MarkupPy==1.14
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.5
protobuf==3.14.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
sqlparse==0.4.1
tablib==2.0.0
uritemplate==3.0.1
urllib3==1.26.2
xlrd==1.2.0
xlwt==1.3.0
